whatsapp-bot/requirements.txt
openai>=1.0.0
python-telegram-bot>=20.0
gspread>=5.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
requests>=2.31.0
python-dotenv>=1.0.0



